requirements.txt
matplotlib==3.4.1
numpy==1.20.2
tensorflow==2.8.0
PySimpleGUI==4.59.0
Pillow==9.0.1

# brew install python-tk